requirements.txt
livekit-agents[openai]~=1.2
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-turn_detector
python-dotenv